moved ActionsCard, AttributeCard and CharacterGeneration into folders and set up containers
